fig/docker: move vendor/ so we don't mount over it
This was very silly of course: building kafka and zookeeper, and then
mounting the development tree without the build dirs over it.

Having fixed that, a lot of tests still fail (but now for other reasons
than that they cannot reach zookeeper, at least).  When we restrict
ourselves to 'make test' (which as of a recent commit only includes the
tests for kafka.pykafka), it's all great.  But: none of the tests there
actually needed us to build kafka or zookeeper...

Oh well.  In any case, tests can run containerised with

    fig run --rm test

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>