utils.socket: avoid overflow exceptions (#151)
We ran into issues with `recvall_into` because it limited the size of
messages that could be received to `socket_receive_buffer_bytes`, and
some messages (such as MessageSets) could be considerably bigger.

This changes how `recvall_into` reads from the socket: while still
reading at most `socket_receive_buffer_bytes`-sized chunks at a time,
it should now be able to assemble messages of any size.

As a downside, that reverts `BrokerConnection.response` to allocating a
fresh bytearray on every call, but to make up for that, this commit now
avoids the slicing/copying that happened in its return statement, as
well as the copies of `chunk` that happened in `recvall_into`.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>