testing the waters with a librdkafka C extension
This is a quick shot to see how librdkafka performs when wrapped through
the python/C API.  Provided you have a cluster at "kafka:9092" and a
topic called "hello" with plenty of messages on partition 0, something
like this should work:

    $ python setup.py build_ext
    $ ## start interpreter in build dir
    > import rd_kafka
    > rd_kafka.consume()

... which returns a (payload, key, offset) tuple, for now.  On my
system, this achieves roughly 20x the throughput of the earlier cffi
based wrapper.  Details from cProfile suggest that on cffi, we incurred
a very high cost in accessing rd_kafka_message_t members, which is not a
problem here, because we can work with native PythonObjs.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>