rdkafka: pass back last-consumed offset
This should allow the SimpleConsumer's autocommit worker to work the way
it does for the pure-python consumer.  Well, mostly: we still need to
check for and handle librdkafka's error values, to prevent writing a new
last-consumed offset when the message was just "end-of-partition" or so.

The message we return is now a namedtuple, with field names matching
SimpleConsumer's Message type (but with an additional field partition_id
which we needed for offset storage).

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>