tests/rdkafka: reuse test_simpleconsumer
Given that we want RdKafkaSimpleConsumer to be as duck-typedly ducky
with SimpleConsumer as it can get, it makes sense to try and have the
exact same test cases run for both.  This commit introduces the minimal
changes to make that work.

There's more we'd like to test, specifically for the style of
integration between librdkafka and pykafka we've chosen, so eg. in the
resume-test we'd want to know if we actually start consuming at the
correct offsets after resuming.  That's still to come.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>