tests/rdkafka: fix unstopped consumers
Failing to stop these consumers resulted in an unexpected interaction
between test cases: it seems keeping the consumer alive prevents the
test cluster from being flushed correctly, so that it is then dirty when
we enter the next TestCase.  As a result we saw inexplicable failures in
pykafka.test_balancedconsumer even though all that code was untouched.

An alternative fix, which also resolved the test failures, was to rename
the topic in either of the TestCases.  I should probably make that
change too.

I'll refactor the consumer.stop() pattern into a contextmanager in an
upcoming commit.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>