rdkafka/simple_consumer: correct block/timeout behaviour
SimpleConsumer.consume() changed considerably as of 8ef83760, and we
need to adjust the meaning of `block` and `consumer_timeout_ms` for
RdKafkaSimpleConsumer accordingly: `block=False` now means "fall through
asap".  Also, we got it partly wrong regardless of those changes: in
RdKafkaSimpleConsumer, `block=True` would always imply waiting
eternally, which was never the case in SimpleConsumer.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>