test_simpleconsumer: resilience against dirty cluster v2
This improves on 55b4116b by avoiding the use of potentially stale
offset values which we may obtain before actually consuming (resulting
in negative values for `offset_diff` on some runs).

We still test what we mean to test: 1) if I commit offsets and then
fetch them, are they the same? and 2) when I resume with a new consumer,
do the committed values actually end up in my OwnedPartitions correctly?

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>