rdkafka: prevent fetch_offsets() on started consumer
This highlights what is of course always going to be an Achilles' heel:
the pykafka and rdkafka sides each hold their own source-of-truth
regarding the current read-offset.

Also, had to improve stop() so that it can handle repeated calls (which
may now occur in tests, because fetch_offsets() needs a stop() first).

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>