test_producer: prevent stalling, error instead
As discussed in #189 this test would previously stall randomly, waiting
forever until you'd hit ctrl-c.  The change demonstrates that the
problem is test-suite related: we put a timeout on the consumer, and as
a result we get a test error whenever None (== timed out) is returned
from consume(), instead of the stalling.

In log output, we see one of two things whenever the test errors out.
Either we get the wrong consumer offset straight away:

    DEBUG - pykafka.simpleconsumer - Set offset for partition 2 to 0

(where offset -1 is expected, on a fresh topic), or

    INFO - pykafka.simpleconsumer - Resetting offsets for 1 partitions

(which, by default settings, will jump to OffsetType.LATEST, ie past the
 freshly produced message we want to get).  Neither of these should
occur if the test harness hands us a newly created topic, as we expect.

On a side note, this resolves #179 because we now test with a binary
string, and that works (when it doesn't stall).

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>