test_producer: dodge issue with test cluster
As noted in the parent commit, this test will occasionally error out,
when it is handed the wrong offsets by the test cluster.  Given that it
is a producer-test, we don't actually need to be fetching consumer
offsets at all, so removing consumer_group works around this nicely.

We don't need to use a balanced consumer either, so this switches away
from that too (I also checked: if we use the simple consumer, but still
set consumer_group, we still get the error, some of the time).

As a bonus this cuts time on this test considerably.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>