test_simpleconsumer: use random topic name
This avoids a test-cluster quirk where sometimes the first call to
fetch_offsets() seems to end up with offsets stored to the cluster
in a previous test run.  I previously thought cb1ba29e had been
sufficient to deal with this problem, but as it turns out, no:

    Traceback (most recent call last):
      File "/srv/pykafka/tests/pykafka/test_simpleconsumer.py", line 64, in test_offset_resume
        self.assertEquals(offsets_resumed, offsets_committed)
    AssertionError: {0: 99, 1: -1, 2: -1} != {0: 99, 1: -1, 2: 199}

With this commit, I can no longer reproduce that (famous last words).

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>