simpleconsumer: more on reset_offsets retry logic
Use of parts_by_error outside the loop where it gets assigned to, meant
that these expressions would only 'see' errors that occurred for the
last broker in that loop.  The changes in the parent commit enable a
quick fix for this.

I'm afraid this comes without a proper test case to demonstrate the
issue.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>