test_simpleconsumer: more on reset_offsets
I realised the failure noted in the parent commit was actually due to a
confusion on my part, caused by the fact that interfaces like
Topic.latest_available_offsets return the *next* offset you might want
to read, whereas SimpleConsumer.reset_offsets accepts the last offset
you want to mark as read.  This rewrites the tests a bit to make that
explicit.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>