simpleconsumer: sync semaphore when flushing queue
This deals with a problem where a manual reset_offsets() (which flushes
the queues of the affected partitions) followed by a consume() would get
stuck in an infinite loop: the semaphore didn't get decremented during
the flush(), and so the consume() keeps cycling partitions trying to
find messages it thinks are there.

The fix seems kind of inefficient, but I hope that's ok because flush()
shouldn't ever be part of a performance-critical operation.

The accompanying test update is how I hit the problem.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>