rdkafka.simple_consumer: handle reset_offsets
As of 2dc86ce (#194) reset_offsets() is a public method that should be
working correctly on a running consumer.  This commit deals with that by
completely nuking the internal rdkafka consumer, which should be ok
unless someone wanted to call reset_offsets() a lot more than expected.
Tests for this are currently on a separate branch, see #213.

As a bonus, this also made fetch_offsets work for a running consumer.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>