test_simpleconsumer: relax reset_offsets requirements
It turned out the "set an invalid offset" part of the last test was
asking a bit more than RdKafkaSimpleConsumer (see #176) can deliver: it
expected invalid offsets to be reset almost immediately, whereas this
reset, if it happens internally to the rdkafka consumer, only gets
visible to us after we consume() a message.

I think that's ok, for the uses where we do this at all: if we passed an
invalid offset to reset_offsets(), that means we have some external
offset store that held an invalid offset anyway, so the fact that
self.held_offsets may keep reporting that invalid offset back to us
for a bit longer (ie until a message is consumed) doesn't make matters
worse.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>