rdkafka: release GIL before calls into librdkafka
As noted in 478ad6fe, that commit introduced some (rare) deadlocks.
What happened there was this: the logging callback would try to grab the
GIL from some librdkafka background thread, and concurrently we were
calling into librdkafka from the main thread (which held the GIL), and
that call would block waiting for the background thread: deadlock.

The fix here releases the GIL for any "non-trivial" calls into
librdkafka.  We may be releasing it a few times too many, but it is nigh
impossible to track down exactly which calls have a potential to
deadlock, so unless performance suffers terribly we'll go with "better
safe than sorry".

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>