Use RdKafkaSimpleConsumer when available
This moves the rdkafka ImportError checking and usage setting out of
Client, where methods like Topic.get_simple_consumer couldn't get to
because they're unaware of Client (and where this was only because we
had originally intended to create a wholly separate rdkafka-based
version of Cluster etc, as per the now archived cffi-based branch).

I checked that catching ImportError is sufficient to handle both the
situation where one didn't compile the extension module, and the case
where dynamically loading the librdkafka lib fails, so this should be
all we need for auto-import of the rdkafka module.

Upcoming: getting the tests to play nicely with this setup.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>