revert py3 str support in protocol methods
This drops get_bytes() conversions from most methods, instead making it
explicit in the test suite that we expect bytes, not str.

If the diffs seem to revert to uglier formatting in some places, that's
because I used git-difftool/vimdiff, trying to reduce the diff with
current master (01d80484) where possible.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>