rdkafka: py3 compatibility
Besides a few str/bytes issues, this was mostly about getting the module
loading system to work both in py2 and py3.  We've taken the simplest
route for now, which means module loading works in py3 but foregoes most
of the new features on offer through PyModuleDef.

On my system, tox succeeds in all of py27, py34, and pypy.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>