cluster: hold only weakrefs to Topics
This breaks a reference cycle between Cluster and Topic.  As per the
discussion in #257, we didn't want to make Topic._cluster a weakref, as
this would break user code that retains a reference to the Topic only.
Instead, we make TopicDict lazily instantiate Topics (which should have
the added benefit of saving a bunch of space if a cluster has many
topics and partitions), while storing only a weakref to it.

This also moves Cluster._update_topics() into TopicDict, because it
needs to know about the weakref implementation details: if a Topic
hasn't been instantiated, it shouldn't try to update it.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>