cluster: fix topic auto-create (even on fresh cluster)
The grandparent commit (which made TopicDict carry weakrefs) introduced
a bug, namely that topic auto-creation via __missing__() caused an
AttributeError to be raised, because I forgot to update __missing__ for
the new nature of the dict entries.

There was no test coverage to catch this, so tests are added here too.
To make those tests pass reliably, we needed to ensure that topic
auto-creation works even on a fresh cluster.  It turns out that with the
updated version of Cluster._get_metadata (per the grandparent commit),
that was just a one-line change - that is, as a side-effect, we've now
also addressed #175: auto-creation works on a freshly bootstrapped kafka
0.8.2 cluster.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>