travis/tox: dumb down paths
This is a small admission of defeat - given that we've only one
non-standard C lib for the time being, we don't need to be quite so
generic in specifying these paths.  For reasons of expedience, then,
I'm not going to figure out how to expand env vars in tox.ini for now.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>