rdkafka: delivery-reporting through futures (WIP)
This passes tests on py27 and py34, but for now crashes and burns on
pypy: I have a hunch that has to do with our reliance on relatively
long-lived opaque pointers to futures (it may mean that we'll have to
disable support for delivery-futures on pypy, or find a better way -
that's for an other commit).

The previous beginnings of futures-support that I'd already put in place
have been given a rethink: the incref'ing of futures, without storing
the actual reference anywhere was bound to cause ref leaks, so now we
actually stick them in a PySet, from which they are deleted when the
delivery callback runs.  Also, the futures we're emitting are now the
py34-native concurrent.futures.Futures.

Still to come: tests, and some workarounds for pypy.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>