rdkafka: fix delivery-reporting on pypy (WIP)
These are the minimal changes to make things work on pypy.  The problem
with the previous implementation was that we'd pass raw PyObject
pointers as opaque pointers through librdkafka, when in fact pypy might
move objects around in memory, so we'd end up with invalid pointers.  As
a workaround, we now store Futures in a dict, using a simple size_t
counter to generate dict keys.  It's this size_t key that we pass around
as the msg_opaque.

In addition, we changed the overall application opaque pointer too: we
previously used an RdkHandle pointer, of which pypy might not be aware
that we were shipping its pointer around (and thus another way to fix
that would have been to incref that object), but now we use the direct
pointer to the dict holding the Futures - which we already held an
explicit reference to, and which should have been a better choice to
begin with.

Still upcoming: we need to do some refactoring (marked with TODOs) to
create a simpler references-cleanup flow.  Also, some comments to
explain the pypy-trickery.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>