producer: avoid messages stuck in queues on stop
Because stop() called _wait_all() after stopping the OwnedBroker threads
(and because the OwnedBroker threads stop as soon as they see
 self.running=False, without trying to check for remaining messages), a
situation could occur where the OwnedBroker threads had already exited
but a message was still sitting in the queue - resulting in _wait_all()
and thus stop() looping forever.  We ended up with occasional hangs in
test_producer.test_async_produce_context.

This is a minimal fix, that just swaps those steps.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>