hold a weak reference to topic in partition
this fixes a hanging broker connection when KafkaClient is deleted, per #257.

This makes it impossible to do something like
client.topics[mytopic].partitions[0].latest_available_offset()
it will fail with a weakref error. This operation can still be performed by simply saving a reference to the
Topic instance in client code prior to calling a method on a Partition instance.

This is slightly inelegant in that it limits the kinds of code clients can write, but it
fixes an important client cleanup issue in a lightweight way.