tests/rdkafka: don't demand zero thread count
This avoids a test failure on pypy, where, if the rdkafka producer-tests
run before the consumer-tests that this commit alters, the latter get
confused because the former's threads haven't disappeared yet.

I thought about forcing the thread clean-up, by making that happen in a
stop() method rather than only when the destructor runs, and then making
sure all tests always call stop() (this is the route that the
simple_consumer tests have taken).  But the code should also work if
users don't call stop(), and so it's better that not all tests call
stop() either.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>