balancedconsumer: stop zk watches if consumer stopped
This partly addresses #276: if we do not own the KazooClient instance,
then upon stopping we'd want to remove from it any watches belonging to
this BalancedConsumer instance.

The solution here is less than ideal, but after a brief study of
ChildrenWatch internals, there seems to be no other way to stop these
watches.  This however only disables the callbacks, but does not really
remove the watches from kazoo (unless I misunderstand the ChildrenWatch
code), and potentially retains references to the BalancedConsumer
instance.  That is, long-term we might want to get rid of this
ChildrenWatch recipe and use a plain kazoo watch.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>