balancedconsumer: always check zk for partitions
This changes _rebalance() to always use zk as the "source of truth" for
what partitions we hold, rather than self._partitions.  The main reason
to do so is that self._partitions may be out of date, if _rebalance() is
triggered by a zk session expiration.  This was originally reported by
@fortime in #215 - I'm however opting for a slightly more conservative
(and thus also less performant!) solution than was proposed there.

Because we now no longer use self._partitions as a proxy for our
zk-registered partitions, I've redefined it to point to the internal
SimpleConsumer's partitions, which simplifies things a bit.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>