balancedconsumer: default param on _setup_internal_consumer
This allows us to revert ugly changes in the tests.  There's a small
change to SimpleConsumer as well, as it turned out that passing an empty
list to it would cause it to grab all partitions, and although this
corner case should never occur, we don't want to accidentally have a
BalancedConsumer grab all partitions if we meant none.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>