balancedconsumer: prevent race on stop
This prevents a race condition where we remove our zk nodes correctly on
stop(), but a concurrently running _rebalance() discovers them missing
and re-registers them.

This should have been added as part of #304 - we didn't detect the issue
there, because the test added with that pullreq only created one rather
than two BalancedConsumers, and the second one is needed to trigger
rebalances.  The test introduced with 8faf6c15e5 reveals the problem by
hanging indefinitely and uninterruptably, iff the tests are run without
a pre-existing cluster (ie the test starts the cluster via get_cluster).

It is not clear how the re-addition of nodes after stop() causes the
test to hang, which is rather frustrating, but may be a hint as to
what's going on in issue #260.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>