balancedconsumer: stop _consumer earlier in _rebalance
Previously, the internal consumer would be left running until the very
end of the rebalancing process, which would potentially stretch out over
several retries.  With these changes, we stop the internal consumer as
soon as it becomes clear that it no longer holds the right set of
partitions, and only then start further zookeeper back-and-forths.

This also removes the offset-commit call, which as of a15b369c is taken
care of by the internal consumer's stop() method.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>