connection: ensure reading the full Size field
BrokerConnection.response() did not check that it received all four
bytes of the Size field that precedes a ResponseMessage.  The result of
reading fewer bytes would be 1) an exception being raised when trying to
`struct.unpack()` it, and 2) the next caller of response() receiving
bogus data (because the previous call didn't `recv` the rest of its
message).

This never tripped up any tests, as it would likely only ever fail on a
heavily loaded network.  It is a good candidate for explaining the kind
of problems reported in issues #318 and #312.

This commit only fixes our reading of the Size field.  Arguably, a more
complete fix would also make sure that the handler, when catching
exceptions from response(), would discard the connection, to prevent
the ensuing bogus data reads.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>