protocol: define Message.__slots__
As per discussion on #269.  In order for this to work, parent classes
also needed to be given `__slots__` (otherwise instances would still
have a `__dict__`).

A few tests used the instance's `__dict__` directly and thus needed a
tweak.  While working that out, I found that `test_snappy_decompression`
wasn't running, because it had been misnamed.  The test wasn't even
py3-compatible yet, but luckily the code it tests was.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>