rdkafka: protect handles with a pthreads shared lock
This has turned into a bit of a monster patch.  The problem it addresses
is that, because we routinely release the GIL in the C extension, we
could not guarantee that accessing rd_kafka_t handles and the like was
ever safe: a concurrent call to stop() from another thread might set
these to NULL at any time.  This is something that in practice only
affected the consumer, in particular when used by a BalancedConsumer
(where zk watches might at any time spawn a rebalancing thread which
might call stop()), but in theory the issue applies to the producer as
well.

The shared lock (which introduces a direct dependency on pthreads, but
not really a new dependency since librdkafka uses pthreads) is set up so
that start() and stop() require exclusive locks, and all other methods
take out a shared lock.  This makes it all a lot easier to reason about:
you grab a lock, and once you have it you only have to check once that
the handle is not NULL.

The performance penalty of locking, at least in a quick informal test,
seems negligible.

I've left one TODO, which concerns the delivery callback (which
currently might try to read a NULL futures dict without checking,
although in practice it will never be NULL), because the changes to be
merged from master (which now features producer futures too) will mean
this is all going to change a lot.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>