rdkafka/consumer: don't wait for the poller thread
We previously waited for the poller thread to join() as part of stop(),
but that was only necessary to prevent segfaults where we tried to
poll the rdkafka handle after stop() had finished.  However, now that
the handle is protected by a pthreads rwlock, this is no longer a
concern and we can save a lot of time by not waiting.

Interestingly, the long wait during stop() caused a lot of odd errors in
BalancedConsumer integration tests, so we might want to plug this join()
call back in when we're on the hunt for timing issues in rebalancing.
One example was that it hit this[0] condition, which erroneously raises
an exception while we're just rebalancing (no need to file a bug as
we're backing that commit out in another pullreq soon).

[0]: https://github.com/Parsely/pykafka/commit/a75df8c5004e5c6310e57593eda06460a63f8e11#diff-c29dae58fe4a50f89318a4d8cf5bfc38R648

I'm also adding anoter pypy skipif, which I had omitted somehow.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>