tests: workaround for #352
I changed this test a few commits back because it had timing issues for
the rdkafka producer, where the message would still be pending and the
consume() would hit a timeout.

The fix however turns out to regularly trip up the non-rdkafka producer,
which will hang indefinitely when it goes wrong (see #352).

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>