tests: avoid race condition in balanced consumer test
This is a hackish fix, because we have no direct way to check if a
rebalancing operation has completed, but should work almost everywhere.
I suspect this fixes #260: the failure was relatively rare with a
pure-python consumer, but because the rdkafka consumer is somewhat
slower in setup/teardown, rebalancing takes longer and we would
regularly see this hang - the issue being that the producer would have
sent its messages before the consumers had rebalanced, and so they would
block forever waiting for messages.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>