producer: replace futures interface by queue
The main motivation for this change was that the futures implementation
was just too resource-intensive, both in time and memory consumption.
On that point I think it delivers: informally benchmarking this commit
with 10**5 calls to produce(), then 10**5 calls to get_delivery_report()
results in a time of 4.5s on my box, whereas the parent commit with
10**5 calls to produce(), and the results passed to futures.wait()
results in a time of 7.3s.

This is just a first shot, and obviously needs some more work: we need
to test that the thread-local queues work as intended (so that you can
use a single producer instance across multiple threads in say, a web
server, and not get the delivery reports mixed up), we need to make
reporting optional (if users don't need delivery reports we don't want
to fill up a queue).

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>