rdkafka: updated producer delivery-reporting
This received another make-over on master, so we needed to get the
interface matching up again.  The seemingly unrelated changes to the
consumer were made because it was simpler to have application opaque
pointers on both consumer and producer handles than to have to check
before DECREF'ing the opaque pointer.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>