tests: rebalance callback hanging [failure]
This would've seemed a straightforward update, now that the "wait for
rebalancing" test helper was merged to master.  However... this waits
for rebalancing to complete all the way until it times out.  (I changed
the consumer-group name thinking that it might be seeing consumers from
other tests, but that's not fixing it either.)  Committing for further
investigation.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>