tests: drop some rebalancing callback assertions
The assertions regarding `held_ids` asserted the opposite of what's
currently true (what you get off the consumer instance are the old
offsets/partitions), and as a result the callback would raise an
`AssertionError` and the rebalance operation would fail.

I think there's no reason to tie us to any behaviour for `held_offsets`
here - if the user needs either the new or the old set of
partitions/offsets, we're already providing those - so it probably makes
sense to drop this from the tests altogether.  Instead, I've updated the
docstring to warn the user about this.

This issue was masked by the `wait_for_rebalancing()` call timing out,
which caused confusion about the real cause of the test failure.  The
test helper now polls for background-thread failures.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>