error_handlers: don't hide KeyErrors
Previously, this utility would silently drop any errors for which we
haven't defined a handler.  The changes here drop the `if errcode in
error_handlers`, so that a KeyError will be raised if we encounter
unexpected errors.  This should help in our Kafka-0.9 compatibility
push, by not concealing any newly defined Kafka 0.9 error codes.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>