simpleconsumer: avoid IllegalGenerationCode on kafka 0.9
This changes the ConsumerGroupGenerationId that we pass with
OffsetCommitRequests.  This field was present in the v2 request format
of kafka 0.8.2, but it wasn't made clear what meaning it had[0].  Trying
to start a SimpleConsumer against a 0.9 cluster however now failed with
an error code 22, which is an IllegalGenerationCode according to the
current protocol docs.

I had no idea what value would satisfy the cluster but found -1 in a
logdump on an old mailing-list thread; it turns out that works.

[0]: https://cwiki.apache.org/confluence/pages/viewpage.action?pageId=61329501

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>