tests: use uuid topic-name in test_balancedconsumer
Although the parent commit claimed that it caused no testing issues, it
turns out it did cause test_consume_latest() to hang a lot.  On log
inspection, and querying the topic offsets from a separate client, it
turns out that during the test, the consumers were assigned offsets that
were ahead of those that the producer in the test produced to, so that
the consume() call would block forever.

This change randomises the topic name for this test, so that offsets
stored in a previous test run don't confuse things any longer.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>