[WIP] benchmarking tools rewrite
This adds a producer benchmark, leverages the `resource` library to get
an idea of memory usage, and generally cleans up the benchmarking code,
which was previously largely a series of strings because that's what
`timeit` expects.  We didn't really benefit from using `timeit` so we do
without it now.

Still coming: updated analysis scripts.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>