client: enable ssl option [WIP]
This is the simplest change that works (except for the rdkafka module,
which still needs updating).  For connecting with the testinstances
cluster (which was configured for TLS in the parent commit), you can do:

    import pykafka, ssl, functools
    ssl_context = ssl.create_default_context(
        cafile="/tmp/kafka-bin/test-certs/root.cert")
    ssl_context.load_cert_chain(
        certfile="/tmp/kafka-bin/test-certs/client.cert",
        keyfile="/tmp/kafka-bin/test-certs/client.key",
        password=b"dWJTxlilbJcH1Wgn")
    wrap = functools.partial(
        ssl_context.wrap_socket,
        server_hostname="localhost")
    client = pykafka.KafkaClient(
        "localhost:9093",
        ssl_wrap_socket=wrap)

However we probably still need to change the interface a bit: we'll need
to store cafile, certfile, keyfile and password, because in the current
format there's no way to obtain these and pass them into the rdkafka
config.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>