cluster: drop ssl-related comment
After more poking around, it turns out that what this "FIXME" wanted
is often not really achievable: if eg. you try sending plaintext
requests to an SSL port on the broker, all you'll see is that
`socket.recv()` et al return null chunks - ie this error will be
indistinguishable from other connection issues.

The particular issue noted, ie that we may end up stuck in "zookeeper
fallback" connection attempts for quite a while, we'll defer to #452
which proposes to remove that fallback altogether.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>