tests: fix test_recover_disconnected()
This test was regularly failing recently.  It has been modified in #513,
which changed it to have both a running consumer and producer at the
point where the brokers are disconnected from.

The result is that it may be the consumer instead of the producer which
attempts connection recovery (problematic because this is the feature
under test), which probably explains the consumer timeouts (consume
calls returning None) observed in test runs.

This fix is the simplest solution, dropping the consume call completely.

Signed-off-by: Yung-Chin Oei <yungchin@yungchin.nl>