Prevent attempts to reset the offset when trying to fetch consumer lag.
`kafka_tools.fetch_consumer_lag` will attempt to reset the offset for a consumer group that does not have any offsets for a topic. This is unexpected behavior, since it seems to indicate it will just read data. Additionally, if the reset fails, it can cause a timeout for the calling service (since consumer_timeout_ms is not set), which can be a problem for monitoring tools.