Include tests in the sdsit but don't install them
By adding the test folder to the manifest but excluding them on
`find_packages`, they will be part of the `sdist`, allowing repackaging
tools to run the tests but they won't be installed when an user pip
installs this package.
This is to prevent the creation of a `tests` module in the user
installation.